+++ src/components/molecules/headers/leftHeaders/LeftBackHeader.tsx
@@ -1,4 +1,4 @@
-import IconButton from '@components/atoms/IconButton';
+import IconButton from '@components/atoms/button/IconButton';
 import colors from '@constants/colors';
 import { useHistory } from 'react-router-dom';
