+++ src/components/molecules/headers/rightHeaders/RightHeader.tsx
@@ -1,4 +1,4 @@
-import IconButton from '@components/atoms/IconButton';
+import IconButton from '@components/atoms/button/IconButton';
 import colors from '@constants/colors';
 
 const RightHeader = () => {
